# api/openapi.yaml
openapi: 3.0.3
info:
  title: Rowbird API
  version: 1.0.0
  description: |
    HTTP API of Rowbird, a self-hosted service that runs SQL queries on a schedule and delivers the
    results. This document is the source of truth for the API (ADR-0005): the Go server interfaces
    and the TypeScript client are generated from it.

    Conventions (docs/spec/05-api.md):

    - Application endpoints live under `/api/v1`. Health endpoints live under `/health`.
    - IDs are UUIDv7 strings. Times are RFC 3339 in UTC.
    - Lists use cursor pagination: `?limit=50&cursor=...` returns `{ "items": [...], "next_cursor": "..." }`.
    - Errors use RFC 9457 `application/problem+json` with a stable `code` and an `i18n_key`.
    - Updates send the current `version`; a mismatch returns `409` with `code: conflict.version`.
    - Secrets are never returned. Secret fields are shown as `{ "configured": true }`.
    - State-changing requests must send `Content-Type: application/json`. Requests authenticated
      with the session cookie must also send `X-CSRF-Token` with the value of the `rowbird_csrf`
      cookie.

    Authorization is declared on each operation with `x-rowbird-authz` (ADR-0017):

    - `role`: minimum workspace role (`viewer` < `editor` < `admin`).
    - `scope`: minimum API key scope (`read` < `run` < `write` < `admin`). Sessions hold every scope.
    - `session_only`: API keys cannot call the operation (account self-service).
...

...
  license:
    name: Apache-2.0
...
servers:
...
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /api/v1/setup:
    post:
      operationId: runSetup
      tags: [setup]
      summary: Create the first admin and sign in
      description: |
        Creates the default workspace, the first admin and the initial settings, then starts a
        session (see `login`). Answers `409 setup.completed` once any user exists.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupRequest"
      responses:
        "201":
          $ref: "#/components/responses/SignedIn"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/v1/auth/login/second-factor:
    post:
      operationId: loginSecondFactor
      tags: [auth]
      summary: Finish signing in with a TOTP code or a recovery code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecondFactorRequest"
      responses:
        "200":
          $ref: "#/components/responses/SignedIn"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/auth/login/second-factor/passkey:
    post:
      operationId: beginPasskeySecondFactor
      tags: [auth]
      summary: Start the passkey step of a sign-in
      description: |
        For a login answered with `mfa_required` whose `methods` include `passkey`. Returns the
        options for `navigator.credentials.get`; the browser's answer goes to `loginSecondFactor`
        as `passkey`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeTokenRequest"
      responses:
        "200":
          description: The WebAuthn request options.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasskeyOptions"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/auth/passkey/options:
    post:
      operationId: beginPasskeyLogin
      tags: [auth]
      summary: Start signing in with a passkey, without a password
      description: Returns the options for `navigator.credentials.get` and a token for `passkeyLogin`. Needs `ROWBIRD_BASE_URL`.
      security: []
      responses:
        "200":
          description: The ceremony.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasskeyCeremony"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/auth/passkey:
    post:
      operationId: passkeyLogin
      tags: [auth]
      summary: Finish signing in with a passkey
      description: |
        The passkey verified the user (PIN or biometrics), so it satisfies the workspace's
        two-factor requirement. Sets the session cookies like `login`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasskeyAnswer"
      responses:
        "200":
          $ref: "#/components/responses/SignedIn"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/v1/auth/password-reset:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Email a link to choose a new password
      description: |
        Always answers 202 with no body, whether or not the email belongs to an account, so it does
        not reveal who has one. When the account exists, is active and has a password, and the
        workspace has a system mailer and `ROWBIRD_BASE_URL`, a one-use link valid for 30 minutes is
        emailed (at most once every 5 minutes per account; a new request replaces earlier links).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "202":
          description: Accepted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/auth/password-reset/confirm:
    post:
      operationId: confirmPasswordReset
      tags: [auth]
      summary: Choose a new password with an emailed link
      description: |
        Sets the password, unlocks the account and signs the user out everywhere. It does not sign
        in; a second factor is still asked at the next sign-in. `410 auth.reset_invalid` for an
        unknown, used or expired link.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetConfirm"
      responses:
        "204":
          description: The password was changed.
        "400":
          $ref: "#/components/responses/BadRequest"
        "410":
          $ref: "#/components/responses/Gone"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: rowbird_session
      description: |
        Browser session. State-changing requests must also send the `X-CSRF-Token` header
        (see `csrfToken`).
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Double-submit CSRF token required with `sessionCookie` on state-changing requests.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: rbk_<random>
      description: API key with scopes `read`, `run`, `write` or `admin`.
...
  responses:
    SignedIn:
      description: Signed in. Sets the `rowbird_session` and `rowbird_csrf` cookies.
      headers:
        Set-Cookie:
          description: Session and CSRF cookies.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Me"
    TooManyRequests:
      description: Too many attempts. Retry after the number of seconds in `Retry-After`.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Gone:
      description: The resource existed but is no longer available.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    BadRequest:
      description: The request is invalid. `errors` lists the offending fields.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Authentication is required.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: The caller is authenticated but not allowed to do this.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Conflict:
      description: The resource changed since it was read (`conflict.version`) or a unique value is taken.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details, extended with a stable code and an i18n key.
      required: [type, title, status, code, i18n_key]
      properties:
        type:
          type: string
          description: URI identifying the problem type, `https://rowbird.dev/errors/<code>`.
          example: https://rowbird.dev/errors/connection.auth_failed
        title:
          type: string
          description: Short human-readable summary in the request locale.
          example: Authentication failed
        status:
          type: integer
          description: HTTP status code.
          example: 422
        code:
          type: string
          description: Stable machine-readable error code.
          example: connection.auth_failed
        i18n_key:
          type: string
          description: Translation key clients can use to show their own message.
          example: errors.connection.auth_failed
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          description: Request path that produced the problem.
        request_id:
          type: string
          description: Id of the request, also sent in the `X-Request-ID` header.
        errors:
          type: array
          description: Field-level validation errors.
          items:
            $ref: "#/components/schemas/FieldError"
        dependents:
          type: array
          description: What uses the resource, for "in use" conflicts.
          items:
            $ref: "#/components/schemas/Dependent"
    FieldError:
      type: object
      required: [field, code]
      properties:
        field:
          type: string
          description: Path of the offending field, for example `port` or `deliveries[0].mode`.
          example: port
        code:
          type: string
          description: Stable validation code.
          example: validation.range
...
    Role:
      type: string
      enum: [viewer, editor, admin]
...
    Locale:
      type: string
      enum: [en, pt-BR]
    Theme:
      type: string
      enum: [system, light, dark]
    Restriction:
      type: string
      enum: [none, password_change_required, mfa_setup_required]
      description: What the session must do before it can use the rest of the API.
...
    SetupRequest:
      type: object
      required: [email, name, password, locale, timezone]
      properties:
        token:
          type: string
        email:
          type: string
        name:
          type: string
        password:
          type: string
          format: password
        locale:
          $ref: "#/components/schemas/Locale"
        timezone:
          type: string
          description: IANA time zone, for example `America/Sao_Paulo`.
    PasswordResetRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
    PasswordResetConfirm:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
          description: The `rbr_` token from the emailed link.
        password:
          type: string
          format: password
...
    SecondFactorRequest:
      type: object
      required: [challenge_token]
      description: Either `code` or `passkey`.
      properties:
        challenge_token:
          type: string
        code:
          type: string
          description: A six-digit TOTP code or a recovery code.
        passkey:
          type: object
          additionalProperties: true
          description: The answer of `navigator.credentials.get` to `beginPasskeySecondFactor`, as JSON.
    ChallengeTokenRequest:
      type: object
      required: [challenge_token]
      properties:
        challenge_token:
          type: string
    PasskeyOptions:
      type: object
      required: [options]
      properties:
        options:
          type: object
          additionalProperties: true
          description: WebAuthn options (the `publicKey` member), with binary values in base64url.
    PasskeyCeremony:
      type: object
      required: [challenge_token, options]
      properties:
        challenge_token:
          type: string
          description: Ties the browser's answer to this ceremony. Valid for 5 minutes, once.
        options:
          type: object
          additionalProperties: true
          description: WebAuthn options (the `publicKey` member), with binary values in base64url.
    PasskeyAnswer:
      type: object
      required: [challenge_token, credential]
      properties:
        challenge_token:
          type: string
        credential:
          type: object
          additionalProperties: true
          description: The answer of `navigator.credentials.get`, as JSON.
...
    Me:
      type: object
      required: [id, email, name, locale, theme, role, totp_enabled, has_password, restriction, version, created_at]
      properties:
        has_password:
          type: boolean
          description: False for accounts that only sign in through OIDC; they skip password confirmations.
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        locale:
          $ref: "#/components/schemas/Locale"
        theme:
          $ref: "#/components/schemas/Theme"
        role:
          $ref: "#/components/schemas/Role"
        totp_enabled:
          type: boolean
        restriction:
          $ref: "#/components/schemas/Restriction"
        last_login_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
...
    Dependent:
      type: object
      required: [type, id, name]
      properties:
        type:
          type: string
          example: query
        id:
          type: string
          format: uuid
        name:
          type: string
...